pyproject.toml
[tool.poetry]
name = "python-ts-graphql-demo"
version = "0.1.0"
description = ""
authors = ["Jökull Sólberg Auðunsson <jokull@solberg.is>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.85.1"
starlette = "^0.16.0"
uvicorn = "^0.15.0"
prisma = "^0.4.3"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
isort = "^5.10.0"
bandit = "^1.7.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 99

[tool.mypy]
plugins = "strawberry.ext.mypy_plugin"
implicit_reexport = false
